Project packages restructed and creation of a controller api for user Project package restruct and creation of a controller api user for entity
It was made a restruction to group classes with the same functionalites,
and it was created a repository, service and controller class for user
entity
to test if the project setup was correct.
The application.yml files was update to include the setup for hibernate
not initialize a h2 database using entity models.