Dependency update
It was removed the jackson dependency to json, because this depenency
it is part of spring web starter.
It was add the dependency of jackson dataformat xml to implement content
negotiation, this dependency is comment because browser set xml as
defaut over json.
It was added the dependency os hateos, but it is commented because needs
to configure a bean to implement.